requirements.txt
paho-mqtt~=2.1.0
loguru~=0.7.3
python-dotenv~=1.1.0
crypto~=1.4.1
pycryptodome~=3.22.0
cryptography~=44.0.2